Add `scrollToSection` to ListActions and ListView
This function scrolls to the section with the provided identifier, with the provided positioning.
The list will first attempt to scroll to the section header. If no header is present, it'll scroll to the first item instead.

If neither header nor items are found, the list will fallback to the footer.

Also, renamed the `firstIndexPath` function in Content to differentiate between querying with an itemIdentifier and a SectionIdentifier.