Updates per PR review
This commit adds `SectionPosition`. `ScrollToSection` will now look for the header or first item if `SectionPosition.top` is specified.
If `.bottom` is specified, the list will scroll to the footer or last section item instead.

Also added `preparePresentationStateForScrollToSection` which checks if the requested section has been loaded. If not, it'll update the presentation state with the index path for that section.

Finally, added some logic to prevent a section footer from appearing underneath a sticky header when scrolled to.