Make `TableListLayout` respect padding by default
Previously while you could set padding on the `TableListLayout`, that
padding would be equally distributed to the left and right rather than
respecting any difference between the two. Even if a custom width was
set on the section or items, if the center alignment was used the
resulting attributes would be globally centered versus centered in the
available width between the padding.

This change resolves the issue on two fronts. First, the
`CustomWidth.Alignment.center` now center aligns between the provided
padding. Second, because the `CustomWidth.default` will center align
completely ignoring padding, this change introduces a root `CustomWidth`
during the layout which all other components merge against to calculate
their origin. This allows the correct center alignment with respect to
padding to cascade down throughout the attributes until a component
overrides the width with something else.