Fix for crash when doing a programmatic scroll down while another external update is processed (#557)
This fixes an issue with multiple updates coming in close to each other
and accidentally "undoing" the programmatic scroll down event that was
previously enqueued. The fix is to instead of just referencing the first
visible index path when processing further updates, to use the
in-progress index path from the prior programmatic scroll event.

I've added a test that fails without the fix.

https://block.atlassian.net/browse/UI-8010