Update the first responder resignation workaround to be enabled by default and cap at < iOS 19 [UI-8849] (#578)
This makes a few changes related to the first responder resignation
workaround that exists for collection views.

- If the `Listable.EnableIOS164FirstResponderWorkaround` user defaults
flag is not set, it will now default to `true`.
- We cap the workaround to OS versions less than iOS 19.0 until the fix
has been verified as required and functioning in the betas that will
come out next week.
- Updated the SystemFlowLayout demo to include the updated issue repro
steps for iOS 17+

### Checklist

Please do the following before merging:

- [x] Ensure any public-facing changes are reflected in the
[changelog](https://github.com/square/Listable/blob/main/CHANGELOG.md).
Include them in the `Main` section.