Reset scroll position if list identifier changes (#588)
If a List’s identifier changes, we should automatically scroll to the
top because maintaining scroll position from one list to another doesn’t
make sense if the identifier (and thus list itself) completely changes.

Reseting the scroll position to the top in this scenario should be done
without animation.

https://block.atlassian.net/browse/UI-8965