fix: Ensure we're referencing up-to-date swipe state when configuring interaction enabled (#595)
For https://block.atlassian.net/browse/UI-9091

This resolves an issue where a cell would not respond to taps after the
"full swipe" swipe action trigger.

The issue was that we were referencing stale swipe state when
configuring `isUserInteractionEnabled`. It was quite unintuitive when
debugging this, but the line `swipeState = state` immediately resulted
in those variables holding different values while processing the swipe
action 🙃. We would incorrectly fall into the
`.willPerformFirstActionAutomatically` case, making the content view
unresponsive, despite the `swipeState` reporting as `.closed`.

The SwipeActionsViewController has also been updated to demonstrate this
is now working as expected.

### Checklist

Please do the following before merging:

- [ ] Ensure any public-facing changes are reflected in the
[changelog](https://github.com/square/Listable/blob/main/CHANGELOG.md).
Include them in the `Main` section.